refactor(login-screen): переносит логику компонента в хук
